Updated HatMe - Changes Galore!
HatMe now uses the component architecture.
In addition to this, the HatMe code has undergone major
restructuring, which resulted in the splitting of it into three
classes. One is the component class, and contains methods for
handling commands and lodaing/unloading the component.
The other two manage permission/restriction checks and the actual
placement of the ItemStacks within the inventory. They are
HatPermsHandler and HatExecutor, respectively.